% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.search.R
\name{explorer_list_tables}
\alias{explorer_list_tables}
\title{List available tables from a given GA4GH search API endpoint.}
\usage{
explorer_list_tables(url)
}
\arguments{
\item{url}{The base URL of the GA4GH Search API Implementation.}
}
\value{
A data frame with the table name and data model, or reference to data model.
}
\description{
List available tables from a given GA4GH search API endpoint.
}
